Moved lenovo lighting
